Changed the folders hierarchy to allow easer REST implementation
Views folder contains what the client 'views'.
layout contains the generic boilerplate to apply to all website pages, can be changed using DOM manipulation.
To use the boilerplate or you want to add page to teacher or student. simply type:
<% layout('./layouts/boilerplate') %>
   <h1>this is the student home page<h1>
<% layout('./layouts/boilerplate') %>

the h1 is copied as html <body> to the boilerplate.

Included bootstrap in the project.

Added partials folder in views which contains a generic navbar and footer, both included in boilerplate.ejs, this can be changed, we can add student navbar and teacher navbar for example and not a generic footer for all pages.